[CHANGE]  "error" to "responseHandler"
