+++ src/app.ts
@@ -11,7 +11,7 @@ import { verifyAuthentication } from './middleware/authUserMiddleware';
 import loginRoutes from './routes/loginRoute';
 import swaggerUi from 'swagger-ui-express';
 import swaggerFile from './swagger_output.json';
-import { errorResponse } from './utils/error';
+import { errorResponse, successResponse } from './utils/responseHandler';
 
 const app = express();
 const port = process.env.PORT ? Number(process.env.PORT) : 4000;
@@ -30,7 +30,6 @@ app.use('/api/comment', verifyAuthentication, commentRoutes);
 app.use('/api/post', verifyAuthentication, postRoutes);
 app.use('/api/tag', verifyAuthentication, tagRoutes);
 app.use('/api', loginRoutes);
-
 
 app.use(errorResponse);
 app.listen(port, () => {
